requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
dataclasses==0.6
filelock==3.0.12
future==0.18.2
idna==2.10
joblib==0.17.0
numpy==1.19.3
packaging==20.4
pandas==1.1.3
protobuf==3.13.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytorch-crf==0.7.2
pytz==2020.1
regex==2020.10.28
requests==2.24.0
sacremoses==0.0.43
scikit-learn==0.23.2
scipy==1.5.3
sentencepiece==0.1.94
six==1.15.0
threadpoolctl==2.1.0
tokenizers==0.9.2
torch==1.7.0
tqdm==4.51.0
transformers==3.4.0
typing-extensions==3.7.4.3
urllib3==1.25.11
